rewrite `@reactpy/client`
